Preference:Theme: Refactoring gui parts classes
- Move ThemeLabel class as inner class
